requirements.txt
afinn==0.1
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
futures==3.1.1
google-translator==0.2
goslate==1.5.1
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
joblib==1.0.1
jsonfield==3.1.0
lyricsgenius==3.0.1
MarkupSafe==2.0.1
nltk==3.6.2
pytz==2021.1
regex==2021.4.4
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
